contrib/globalsign/mgo: moving context into mongoConfig
